+++ autofit/database/query_model/__init__.py
@@ -2,8 +2,7 @@ from .condition import (
     NameCondition as N,
     ValueCondition as V,
     StringValueCondition as SV,
-    TypeCondition as T,
-    And,
-    Or
+    TypeCondition as T
 )
+from .junction import And, Or
 from .query import NamedQuery as Q
